MRCD8-165 Change dialog alter into plugins
