+++ modules/mrc_media/mrc_media.services.yml
@@ -13,3 +13,10 @@ services:
     arguments:
       - '@entity_type.manager'
       - '@current_user'
+  plugin.manager.media_embed_dialog_manager:
+    class: Drupal\mrc_media\MediaEmbedDialogManager
+    parent: default_plugin_manager
+  plugin.manager.video_option_manager:
+      class: Drupal\mrc_media\VideoOptionManager
+      parent: default_plugin_manager
+
