+++ modules/mrc_media/src/Annotation/MediaEmbedDialog.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Drupal\mrc_media\Annotation;
+
+use Drupal\Component\Annotation\Plugin;
+
+/**
+ * Defines an entity browser display annotation object.
+ *
+ * @see hook_entity_browser_display_info_alter()
+ *
+ * @Annotation
+ */
+class MediaEmbedDialog extends Plugin {
+
+  /**
+   * The plugin ID.
+   *
+   * @var string
+   */
+  public $id;
+
+  /**
+   * The machine name of the media type to alter.
+   *
+   * @ingroup plugin_translatable
+   *
+   * @var \Drupal\Core\Annotation\Translation
+   */
+  public $media_type;
+
+}
